pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "molbench"
version = "0.1.0"
description = "Molecule benchmarks"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "pyscf==2.11.0",
  "matplotlib>=3.7.2",
  "qc-PyCI==0.6.3",
  "block2>=0.5.3",
  "hydra-core==1.3.2",
  "omegaconf==2.3.0",
  "jupyter",
  "mpi4py",
]

[project.scripts]
molbench = "molbench.main:main"

[tool.setuptools]
packages = ["molbench"]
